Match DB types to Java types

+++ src/main/resources/db/migration/V1_0_3__create_products.sql
@@ -1,5 +1,5 @@
 CREATE TABLE product (
-    id integer PRIMARY KEY,
+    id bigserial PRIMARY KEY,
     name varchar(255) UNIQUE NOT NULL,
     description varchar(500),
     icon varchar(255)
