+++ src/main/resources/db/migration/V1_0_4__create_hospitals_needs.sql
@@ -1,7 +1,7 @@
 CREATE TABLE HOSPITAL_NEEDS(
-id serial PRIMARY KEY,
-hospital_id integer NOT NULL,
-product_id integer NOT NULL,
-quantity integer NOT NULL,
+id bigserial PRIMARY KEY,
+hospital_id bigint NOT NULL,
+product_id bigint NOT NULL,
+quantity bigint NOT NULL,
 request_timestamp timestamp NOT NULL default now(),
  FOREIGN KEY(hospital_id) REFERENCES hospital(id));
